Graph class very begining && get-dates-for-news-api renamed to get-request-dates
