+++ src/js/Modules/NewsApi.js
@@ -1,9 +1,9 @@
 import Api from './Api';
 
 export default class NewsApi extends Api {
-  constructor(getDatesForNewsApi) {
+  constructor(getRequestDates) {
     super();
-    this.getDatesForNewsApi = getDatesForNewsApi;
+    this.getRequestDates = getRequestDates;
     this.param = {
       headers: {
         'X-Api-Key': '7c22611ba74c47b9bb7bab94a85a00f9'
@@ -12,7 +12,7 @@ export default class NewsApi extends Api {
   }
 
   async fetchNews(request) {
-    const [fromDate, toDate] = this.getDatesForNewsApi();
+    const [fromDate, toDate] = this.getRequestDates();
     const url = `http://newsapi.org/v2/everything?q=${request}&from=${fromDate}&to=${toDate}&sortBy=publishedAt&language=ru&pageSize=100`;
     return super.fetchData(url, this.param);
   }
