+++ src/js/components/Graph.js
@@ -1,0 +1,15 @@
+export default class Graph {
+  constructor(request, newsData) {
+    this.request = request;
+    this.news = newsData;
+  }
+
+  init() {
+    this.renderMonth();
+  }
+
+  renderMonth() {
+    const graphMontElement = document.querySelector('.graph__title-month');
+    console.log(graphMontElement)
+  }
+}
