+++ src/js/statistic.js
@@ -5,6 +5,7 @@ import '../scss/styles-page-statistic.scss';
 
 import DataStorage from './Modules/DataStorage';
 import Statistic from './components/Statistic';
+import Graph from './components/Graph';
 
 import getMentionsInTitles from './utils/get-mentions-in-titles';
 
@@ -14,5 +15,7 @@ if (sessionStorage.newsData) {
   const request = dataStorage.getData('request');
   const newsData = dataStorage.getData('newsData');
   const statistic = new Statistic(request, newsData, getMentionsInTitles);
+  const graph = new Graph(request, newsData);
   statistic.init();
+  graph.init();
 }
